refactor(controllers): extend playing writer class from Writer
